+++ backend/src/main/java/com/nashtech/ecommercialwebsite/dto/response/JwtResponse.java
@@ -5,11 +5,11 @@ import java.util.List;
 public class JwtResponse {
     private String token;
     private String type = "Bearer";
-    private Long id;
+    private Integer id;
     private String username;
     private List<String> roles;
 
-    public JwtResponse(String accessToken, Long id, String username, List<String> roles) {
+    public JwtResponse(String accessToken, Integer id, String username, List<String> roles) {
         this.token = accessToken;
         this.id = id;
         this.username = username;
@@ -32,11 +32,11 @@ public class JwtResponse {
         this.type = tokenType;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
